### 2017-04-21 #245 @swillner
coauthors: @rgieseke, @cahartin
