### 2017-04-21 4f86642
fixed spinup_size reset
