### 2017-04-21 #245 @sgrieve
I'd be happy to review. I should be able to work on it early next week at the latest.
