### 2017-04-24 454435b
Test disabling spin-up
