### 2017-04-24 a18a9f4
Test in_spinup flag for observables
