### 2017-04-24 d8fba8f
PEP8 fixes
