### 2017-04-24 3f702b0
More documentation for observables
Also improves formatting
