### 2017-04-24 7bbd97d
Update for v0.6.0
[skip ci]
