### 2017-04-24 #245 @rgieseke
Thanks @sgrieve!

We've just updated Pyhector to v0.6.0. Is there a `whedon` command to change/regenerate the metadata?
