### 2017-04-24 #245 @arfon
> I think that might be a question for @arfon ...

👍 it's a manual process right now (I've just done it)
