### 2017-04-24 #248 @whedon
[REVIEW]: pyhector: A Python interface for the simple climate model Hector
**Submitting author:** @swillner (<a href="http://orcid.org/0000-0001-6798-6247">Sven N. Willner</a>)
**Repository:** <a href="https://github.com/openclimatedata/pyhector/" target ="_blank">https://github.com/openclimatedata/pyhector/</a>
**Version:** v0.7.0
