### 2017-04-24 #245 @arfon
@sgrieve - please head over to #248 to actually carry out the review
