### 2017-04-24 #248 @sgrieve
I have updated the version number in the review comment, as it didn't copy across from the pre-review. I'm hoping to have time to work on this review tomorrow. I will post any questions I have here, and if I find any bugs I will open issues on the source repo.
