### 2017-04-24 960ef90
Amend OS X installation note
User reported problems when using a system provided Python version.
See #17.

[skip ci]
