### 2017-04-25 #248 @sgrieve
I have been working on this today and overall I am very happy with the quality of the code and documentation. I'd like to thank the authors for this contribution, its been a very enjoyable day working through this submission (much better than marking exams!).

I have checked all of the items on the checklist and so the minor issues I raise below are not severe, but are more suggestions that I think may improve the paper and the code.

* In the readme, the usage guide is placed before the installation information. It might flow better to have the install guide before the usage section.
* When running the example code and some of the code from the jupyter workbook, I had problems because of the use of the unicode degree symbol in the plot y axis labels. I was able to fix this easily, but being thrown an error message right off the bat may discourage some users, so it might be a good idea to avoid using the degree symbol to eliminate that possible barrier to entry.
* I would suggest adding ``plt.show()`` to the last line of the complex example in the readme, if someone runs the code in a non-interactive session no plot will display otherwise.
* When reading the API, it states:

> List of requested output variables as strings. if set to “all” returns all available variables. Defaults to global temperature, CO2 concentration and forcing. A full list is in variables

But I cannot find this full list. I was able to generate it easily from the code to play around with the model, but it would be useful to have this list reproduced somewhere in the documentation.

* In a similar vein, the descriptions in the API for ``pyhector.constants`` through to ``pyhector.units`` are quite sparse. I think it would help a prospective user to see some examples here, if not the full range of value for each of these variables.

* I just want to check about the authorship of the paper, there are 3 authors, yet only 2 have been involved in the development of pyhector based on the commit history of the repo. Have all authors contributed to this submission?

* In the paper, it would be better to cite pandas rather than using a footnote. Information on citing it can be found here: https://www.scipy.org/citing.html#pandas

Once these comments have been addressed I will be delighted to accept this paper.
