### 2017-04-25 c6fe73d
Use McKinney paper as Pandas reference
[skip ci]
