+++ paper/paper.bib
@@ -42,3 +42,11 @@
   doi = {10.1109/MCSE.2007.53},
   publisher = {IEEE Computer Society},
 }
+@InProceedings{McKinney2010,
+  title = {Data Structures for Statistical Computing in Python},
+  author = {Wes McKinney},
+  booktitle = {Proceedings of the 9th Python in Science Conference},
+  pages = {51 - 56},
+  year = 2010,
+  editor = {St\'efan van der Walt and Jarrod Millman}
+}
