### 2017-04-25 eac1c98
Move installation notes to top
