### 2017-04-25 74f6c9c
Ensure plot output in non-interactive session
[skip ci]
