### 2017-04-26 8c98358
Better documentation for sub-modules with config variables
