+++ docs/conf.py
@@ -22,6 +22,17 @@ import sys
 
 sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
 
+from os.path import basename
+
+from sphinx.util.compat import Directive
+from docutils import nodes, statemachine
+
+try:
+    from StringIO import StringIO
+except ImportError:
+    from io import StringIO
+
+
 class Mock(object):
     def __init__(self, *args, **kwargs):
         pass
@@ -41,6 +52,32 @@ class Mock(object):
 MOCK_MODULES = ['numpy', 'pandas']
 for mod_name in MOCK_MODULES:
     sys.modules[mod_name] = Mock()
+
+# From http://stackoverflow.com/a/18143318 by
+# http://stackoverflow.com/users/839411/alex-forencich
+class ExecDirective(Directive):
+    """Execute the specified python code and insert the output into the document"""
+    has_content = True
+
+    def run(self):
+        oldStdout, sys.stdout = sys.stdout, StringIO()
+
+        tab_width = self.options.get('tab-width', self.state.document.settings.tab_width)
+        source = self.state_machine.input_lines.source(self.lineno - self.state_machine.input_offset - 1)
+
+        try:
+            exec('\n'.join(self.content))
+            text = sys.stdout.getvalue()
+            lines = statemachine.string2lines(text, tab_width, convert_whitespace=True)
+            self.state_machine.insert_input(lines, source)
+            return []
+        except Exception:
+            return [nodes.error(None, nodes.paragraph(text = "Unable to execute python code at %s:%d:" % (basename(source), self.lineno)), nodes.paragraph(text = str(sys.exc_info()[1])))]
+        finally:
+            sys.stdout = oldStdout
+
+def setup(app):
+    app.add_directive('exec', ExecDirective)
 
 
 # -- General configuration ------------------------------------------------
