### 2017-04-26 699181c
Fix tests after rename
