### 2017-04-26 55b42c5
updated changelog and fixed docs dependencies
