### 2017-04-26 #248 @rgieseke
Many thanks for the detailed review! Answers to the suggested changes below:

> In the readme, the usage guide is placed before the installation information. It might flow better to have the install guide before the usage section.

Good idea, done.

>    When running the example code and some of the code from the jupyter workbook, I had problems because of the use of the unicode degree symbol in the plot y axis labels. I was able to fix this easily, but being thrown an error message right off the bat may discourage some users, so it might be a good idea to avoid using the degree symbol to eliminate that possible barrier to entry.

A colleague reported just that when testing, so this might be very useful change. Not sure whether there is a nice Python2/3 workaround so going with a plain text unit.

> I would suggest adding plt.show() to the last line of the complex example in the readme, if someone runs the code in a non-interactive session no plot will display otherwise.

Good catch, spoiled by mostly plotting from Notebooks these days!

> List of requested output variables as strings. if set to “all” returns all available variables. Defaults to global temperature, CO2 concentration and forcing. A full list is in variables
> But I cannot find this full list. I was able to generate it easily from the code to play around with the model, but it would be useful to have this list reproduced somewhere in the documentation.
> In a similar vein, the descriptions in the API for pyhector.constants through to pyhector.units are quite sparse. I think it would help a prospective user to see some examples here, if not the full range of value for each of these variables.

Very good point, the configuration dictionaries in the sub-modules have now more detailed documentation in the API-docs section. They are auto-generated so they should stay up-to-date when Hector variables change.

> I just want to check about the authorship of the paper, there are 3 authors, yet only 2 have been involved in the development of pyhector based on the commit history of the repo. Have all authors contributed to this submission?

@cahartin contributed as a Hector developer with explaining and adapting Hector and by ensuring that Hector and Pyhector produce identical results. The Pyhector software requires the hector-wrapper and slightly modified Hector mentioned in the Readme. We therefore think we should archive the PyPI release which includes everything required to build Pyhector.

> In the paper, it would be better to cite pandas rather than using a footnote. Information on citing it can be found here: https://www.scipy.org/citing.html#pandas

Done!
