### 2017-04-26 #248 @swillner
@sgrieve Thank you very much! We have tagged the revised version "v0.7.0". Could you update that again,  please? ;)
