### 2017-04-27 #248 @sgrieve
Great. I've updated the version number and I'm happy that my suggestions have all been addressed. So I am happy to accept the paper :rocket:

I think the next step is for you to generate a doi for the package using zenodo or figshare.

@arfon Is there anything else I need to do?
