### 2017-04-27 #248 @swillner
Cool! We have uploaded the pypi-archive (since pyhector is a multi-repo software) to zenodo:
https://doi.org/10.5281/zenodo.569492
