### 2017-04-28 #248 @arfon
@sgrieve many thanks for your review here ✨

@swillner - your paper is now accepted into JOSS. The DOI is http://dx.doi.org/10.21105/joss.00248 :zap: 🚀 💥
